refactor(server): manager calls rustfs/mailpit directly; drop adapter
Replace the polymorphic services.AllBinary() loop in reconcileBinaryServices
with two explicit blocks (Source 1a/1b) that call rustfsproc.BuildSupervisorProcess()
and mailpitproc.BuildSupervisorProcess() directly — matching the redis/postgres/mysql
pattern already in use.

Introduce internal/rustfs/proc and internal/mailpit/proc as leaf sub-packages
(no internal/server dependency) to break the import cycle that would arise if
internal/server imported the parent packages directly. The parent packages
delegate BuildSupervisorProcess() and Binary() to the sub-packages.

Rename binary_service.go → daemon_status.go; remove buildSupervisorProcess
and buildReadyFunc adapters (now dead code). Delete binary_service_test.go
(tests for deleted adapters and for BuildSupervisorProcess, already covered by
per-tool service_test.go files); keep writeDaemonStatus round-trip test in
daemon_status_test.go.
